moved corsmiddleware and requestlogger functions to pkg/middleware/*.go
