+++ pkg/middleware/cors.go
@@ -1,0 +1,30 @@
+package middleware
+
+import (
+	"fmt"
+	"net/http"
+	"strings"
+
+	"github.com/gin-gonic/gin"
+	"github.com/kriipke/console-api/pkg/initializers"
+	"github.com/kriipke/console-api/pkg/models"
+	"github.com/kriipke/console-api/pkg/utils"
+)
+
+func CORSMiddleware() gin.HandlerFunc {
+    return func(c *gin.Context) {
+		config, _ := initializers.LoadConfig("configs")
+
+		c.Writer.Header().Set("Access-Control-Allow-Origin", config.ClientOrigin)
+        c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
+        c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
+        c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT")
+
+        if c.Request.Method == "OPTIONS" {
+            c.AbortWithStatus(204)
+            return
+        }
+
+        c.Next()
+    }
+}
